#12 Add database

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -3,7 +3,7 @@ package ru.yandex.practicum.filmorate.service;
 import ru.yandex.practicum.filmorate.model.User;
 
 public interface UserService {
-	User get(int userId);
+	User get(long userId);
 
 	User save(User user);
